backend/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1
requests~=2.31.0
bcrypt~=4.1.2
PyJWT~=2.8.0
psycopg2~=2.9.9
Flask_Cors~=4.0.0
python-dotenv~=1.0.1